+++ src/main.tsx
@@ -2,13 +2,17 @@ import './styles.css';
 
 import React from 'react';
 import { createRoot } from 'react-dom/client';
+import { ThemeProvider } from 'styled-components';
 
-import App from './App';
+import HomePage from './pages/HomePage';
+import { theme } from './theme';
 
 const container = document.getElementById('root');
 const root = createRoot(container!);
 root.render(
   <React.StrictMode>
-    <App />
+    <ThemeProvider theme={theme}>
+      <HomePage />
+    </ThemeProvider>
   </React.StrictMode>,
 );
