+++ src/pages/HomePage/index.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import DefaultLayout from '../../layouts/DefaultLayout';
+
+const HomePage: React.FC = () => {
+  return (
+    <DefaultLayout title="Portfolio">
+      <div className="flex flex-col"></div>
+    </DefaultLayout>
+  );
+};
+
+export default HomePage;
