+++ src/shared/components/Icon/components/Email.tsx
