+++ src/shared/components/Icon/components/Github.tsx
