+++ src/shared/components/Icon/components/PlayStore.tsx
