+++ src/shared/components/Icon/components/Web.tsx
