+++ src/shared/components/Introduction/components/Social.tsx
