+++ src/shared/components/Introduction/index.tsx
