+++ src/shared/components/Projects/components/Project.tsx
