+++ src/shared/components/SEO.tsx
